server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Werkzeug==3.0.3
python-dotenv==1.0.1
bcrypt==4.1.3
pymongo==4.7.3
cloudinary==1.36.0